Update README.md
The gui was removed over a year ago and this readme was forgotten to be updated.

I am starting to regret even making open source projects in the first place when half of the internet seems to be filled with the worst people of all times. -_- commit done in response to overly aggressive and non understanding people looking only into finding ways to get into an argument online.